huggingface_space/requirements.txt
gradio==4.16.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
